Move auth0-tokens out of soon to be removed utils/user
