+++ app/utils/user.js
@@ -74,17 +74,3 @@ const getTokenExpires = async () => {
 };
 
 const getCurrentUnixTimestamp = () => Math.floor(Date.now() / 1000);
-
-/**
- * Save the JWT token that Auth0 supplies.
- *
- * This can be used for authentication against our own backend API.
- * When decoded it contains user profile and a unique identifier: "sub"
- *
- * @param {str} token JWT token that Auth0 supplied
- */
-export const setAuth0Token = async token => {
-  saveItem('auth0token', token);
-};
-
-export const getAuth0Token = async () => getItem('auth0token');
